requirements.txt
google-genai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
nemo_toolkit[asr]
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
